Add a master agent and a pulse generator agent
The master agent is the one responsible to spawm a specified
number of neurons at the beginning. It also spawns a pulse
generator agent.

The pulse generator is created for testing purposes. It sends
events with the label "electric pulse". When a growth cone
receives the event, it starts growing in that direction, and
creates new axon segment agents on its way.